Separate panel into presentational and container components
