+++ lib/session.js
@@ -9,9 +9,9 @@ import type { Reducer } from "redux"; // eslint-disable-line
 import Disposables from "./disposables";
 import * as A from "./action-creators.js";
 import NotesFileFilter from "./notes-file-filter";
+import newApp from "./react/containers/app";
 import newRootReducer from "./reducers/index";
 import NotesCache from "./notes-cache";
-import newPanelContainer from "./react/containers/panel";
 import RestartSessionForNewConfigToTakeEffect from "./restart-session-for-new-config-to-take-effect";
 import ToggleAtomWindow from "./toggle-atom-window";
 import TogglePanel from "./toggle-panel";
@@ -86,10 +86,10 @@ export default class Session {
     };
     this._store = createStore(rootReducer, initialState, middlewares);
 
-    const PanelContainer = newPanelContainer(columns);
+    const App = newApp(columns);
     render(
       <Provider store={this._store}>
-        <PanelContainer />
+        <App />
       </Provider>,
       this._panel.getItem()
     );
