+++ styles/textual-velocity.less
@@ -52,7 +52,7 @@
   padding-right: @component-icon-padding;
 }
 
-.tv-items {
+.textual-velocity {
   .header {
     position: relative;
     overflow-y: scroll;
